requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas==2.2.1
openpyxl==3.1.2
pyarrow==15.0.2
sqlalchemy==2.0.30
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
